ci: remove auto committing of generated files
The current system of automatically generating and pushing commits to
feature branches is likely not working as intended. The current system
seem to do a lot of unintended changes when trivial changes to the
source code is applied.

This ensures that no one can merge anything into to the default branch
without having to understand all the generated code changes.

In addition, feature branches should be complete when pushed by a
developer, and should not receive additional commits upon being pushed,
as this would be surprising to many developers.

Also renamed the relevant names in the workflow file.